Refactor volunteer sign up into separate pages
